+++ projects/fusion-docs/src/app/pages/components/input-docs/input-docs.component.ts
@@ -1,6 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {FormBuilder, FormControl, Validators} from '@angular/forms';
-import {InputOptions, InputSize, StyleVersion} from '@ironsource/fusion-ui';
+import {StyleVersion} from '@ironsource/fusion-ui';
+import {InputOptions, InputSize} from '@ironsource/fusion-ui/components/input/common/base';
 import {TooltipPosition} from '@ironsource/fusion-ui/components/tooltip/common/base';
 import {BehaviorSubject} from 'rxjs';
 import {DocsMenuItem} from '../../../components/docs-menu/docs-menu';