+++ projects/fusion-docs/src/app/pages/components/table-docs/table-docs.component.ts
@@ -776,7 +776,7 @@ export class TableDocsComponent implements OnInit, OnDestroy {
             console.log(`row removed`, rowRemoved);
         });
         this.selectedVersion$.subscribe((styleVersion: StyleVersion) => {
-            if (styleVersion === StyleVersion.V2) {
+            if (styleVersion === StyleVersion.V2 || styleVersion === StyleVersion.V3) {
                 this.router.navigate(['docs/components/v2/table']);
             }
         });