+++ projects/fusion-docs/src/app/pages/components/video-player-docs/video-player-docs.component.ts
@@ -67,7 +67,7 @@ export class VideoPlayerDocsComponent implements OnInit {
 
     ngOnInit() {
         this.selectedVersion$.subscribe((styleVersion: StyleVersion) => {
-            if (styleVersion === StyleVersion.V2) {
+            if (styleVersion === StyleVersion.V2 || styleVersion === StyleVersion.V3) {
                 this.router.navigate(['docs/components/v2/video-player']);
             }
         });