+++ projects/fusion-ui/components/button/v2/components/button-loading/button-loading.component.ts
@@ -7,4 +7,5 @@ import {Component, Input} from '@angular/core';
 })
 export class ButtonLoadingComponent {
     @Input() projectContent: boolean = false;
+    @Input() buttonWithIcon: boolean = false;
 }