+++ projects/fusion-ui/components/modal/common/base/public-api.ts
@@ -1,0 +1,2 @@
+export * from './modal.base.service';
+export * from './modal.base.component';