+++ projects/fusion-ui/components/modal/v3/components/modal-content/modal-content.component.ts
@@ -1,0 +1,12 @@
+import {ChangeDetectionStrategy, Component, Input} from '@angular/core';
+
+@Component({
+    selector: 'fusion-modal-content',
+    templateUrl: './modal-content.component.html',
+    styleUrls: ['./modal-content.component.scss'],
+    changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class ModalContentComponent {
+    @Input() padding: string;
+    @Input() waiting = false;
+}